group assemblers in one file
