+++ src/factor_nodes/clamp.jl
@@ -23,9 +23,9 @@ mutable struct Clamp{T<:VariateType} <: DeltaFactor
     value::Any
 
     # Fields for algorithm assembly
-    dist_or_msg::Type
-    buffer_id::Symbol
-    buffer_index::Int64
+    dist_or_msg::Type # Define whether the clamp outputs a Message or ProbabilityDistribution
+    buffer_id::Symbol # Specify the buffer id for an attached placeholder
+    buffer_index::Int64 # Specify the buffer index when the placeholder attaches a vector
 
     function Clamp(out::Variable, value::Any; id=generateId(Clamp{variateType(value)}))
         self = new{variateType(value)}(id, Array{Interface}(undef, 1), Dict{Symbol,Interface}(), value)
@@ -117,21 +117,4 @@ function placeholder(   buffer_id::Symbol;
 
     var = Variable(id=var_id)
     return placeholder(var, buffer_id, index=index, default=default, datatype=datatype, dims=dims)
-end
-
-"""
-Depending on the origin of the Clamp node message, contruct a message or marginal inbound
-"""
-function assembleClamp!(inbound::Clamp, dist_or_msg::Type)
-    inbound.dist_or_msg = dist_or_msg
-    if inbound in keys(ForneyLab.current_graph.placeholders)
-        # Message comes from data buffer
-        (buffer, idx) = ForneyLab.current_graph.placeholders[inbound]
-        inbound.buffer_id = buffer
-        if idx > 0
-            inbound.buffer_index = idx
-        end
-    end
-
-    return inbound
 end
